requirements.txt
Flask
Flask-Limiter
gunicorn
psutil
pymemcache
pynacl
python-jose
python-pam
redis
requests
statsd